# openAPI_specification/OpenAPI.yaml
openapi: "3.0.3"
info:
  title: Dummy Roblox Shop
  version: "1.0"
  description: A Codecademy exercise building for fun very basic online e-store - Roblox Shop API
servers:
  - url: https://api.server.test/v1
...
security:
  - BasicAuth: []
paths:
...
  /products/product/:productId:
    summary: Get product data
    description: This path is used for retreiving a single product data
    get:
      tags:
        - Products
      summary: Get product details
      description: Retreive product data based on product ID
      operationId: get_product
      parameters:
        - $ref: "#/components/parameters/productIdPathParameter"
      responses:
        "200":
          description: Product data retrieved successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /products/allProducts:
    summary: Get all products from DB
    description: This path is used for retreiving all products from DB
    get:
      tags:
        - Products
      summary: Retrieve all products from DB
      description: Retreive all products from DB
      operationId: get_all_products
      responses:
        "200":
          description: All products data successfully retreived from DB
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
  /cart:
    summary: Getting cart
    description: This path is used for retreiving cart details if user is logged in.
    get:
      tags:
        - Cart
      summary: Get user cart
      description: Retreive Cart data for logged in user.
      operationId: get_cart
      responses:
        "200":
          description: Cart data retreived successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  userid:
                    type: integer
                  Products:
                    type: array
                    items:
                      $ref: "#/components/schemas/CartProductWithQty"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /cart/addProduct:
    summary: Add product to Cart
    description: This route is used for adding product to a cart or increasing qty of the product in the cart
    post:
      summary: Add item or Increase Cart item Qty
      description: Add product or increase product qty
      tags:
        - Cart
      operationId: cart_add_product
      requestBody:
        description: Passing product Id
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductId"
      responses:
        "303":
          description: Product added or qty increased successfully, user redirected to Cart
          headers:
            Location:
              $ref: "#/components/headers/Location"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /cart/removeProduct:
    summary: Remove product from Cart based on product ID
    description: This route is used for removal of Product from the Cart
    delete:
      summary: Remove item from Cart
      description: Delete product from Cart
      tags:
        - Cart
      operationId: cart_detete_product
      requestBody:
        description: Passing product Id
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductId"
      responses:
        "303":
          description: Product removed from Cart successfully, user redirected to Cart
          headers:
            Location:
              $ref: "#/components/headers/Location"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /cart/reduceProduct:
    summary: Reduce or Remove product from the Cart
    description: This route is used for reducing qty of product from the Cart or removing the Product from the Cart
    delete:
      summary: Reduce Qty or Remove item from Cart
      description: Reduce qty or remove product from Cart
      operationId: cart_reduce_product
      tags:
        - Cart
      requestBody:
        description: Passing product Id
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductId"
      responses:
        "303":
          description: Product qty reduce or product removed from cart successfully, user redirected to Cart.
          headers:
            Location:
              $ref: "#/components/headers/Location"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /orders:
    summary: Retreive all User orders
    description: This path is used for retreiving all orders of the logged in user
    get:
      summary: Get all orders of User with order details
      description: Get all User Orders
      operationId: get_orders
      tags:
        - Orders
      responses:
        "200":
          description: All orders data retreived successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                    userid:
                      type: integer
                    Products:
                      type: array
                      items:
                        $ref: "#/components/schemas/OrderProductWithQty"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /orders/order/:orderId:
    summary: Get specific orders
    description: This path is used for retreiving specific order based on passed ID
    get:
      summary: Get specific order details
      description: Get specific order based on order Id
      tags:
        - Orders
      operationId: get_order
      parameters:
        - $ref: "#/components/parameters/orderIdPathParameter"
      responses:
        "200":
          description: All orders data retreived successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  userid:
                    type: integer
                  Products:
                    type: array
                    items:
                      $ref: "#/components/schemas/OrderProductWithQty"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /orders/cartToOrder:
    summary: Creates order from Cart items
    description: This path is used for transition of Cart items to actual Order
    post:
      summary: Create Order from Cart
      description: Create order from Cart
      operationId: cart_to_order
      tags:
        - Orders
      responses:
        "200":
          description: Order created successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    description: order ID
                  userid:
                    type: integer
        "500":
          $ref: "#/components/responses/InternalServerError"
...
components:
  parameters:
    productIdPathParameter:
      name: orderId
      description: Order ID number
      in: path
      required: true
      schema:
        type: integer
    orderIdPathParameter:
      name: orderId
      description: Order Id
      in: path
      required: true
      schema:
        type: integer
...
  schemas:
...
    ProductId:
      type: object
      properties:
        productId:
          type: integer
      required:
        - productId
    BaseProduct:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
          format: double
        category:
          type: string
          enum:
            [
              "general",
              "outerwear",
              "pants",
              "skirts",
              "headwear",
              "accessories",
              "shoes",
            ]
      required:
        - name
        - description
        - price
        - category
    Product:
      allOf:
        - type: object
          properties:
            id:
              type: integer
          required: [id]
        - $ref: "#/components/schemas/BaseProduct"
    CartProductWithQty:
      allOf:
        - $ref: "#/components/schemas/Product"
        - type: object
          properties:
            CartsProducts:
              type: object
              properties:
                qty:
                  type: integer
...
    OrderProductWithQty:
      allOf:
        - $ref: "#/components/schemas/Product"
        - type: object
          properties:
            "OrdersProducts":
              type: object
              properties:
                qty:
                  type: integer
                  nullable: false
...
  headers:
    Location:
      description: The redirect URI
      schema:
        type: string
        example: /user/profile
  responses:
    InternalServerError:
      description: Error at server end.
      content:
        application/json:
          schema:
            type: object
            properties:
              msg:
                type: string
                description: Error message
  securitySchemes:
    BasicAuth:
      type: http
      scheme: basic